Develop

+++ src/fetch.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import pickle
 
 import requests
@@ -9,20 +8,25 @@ from src import config
 
 
 class Fetcher:
-    def __init__(self):
+    def __init__(self, username="", password=""):
         self.atcoder_url = "https://atcoder.jp"
         self.config_path = config.PROJECT_ROOT + "/data/config/"
         self.session_filename = "session.pkl"
-        self.is_login = True
-        if not self.__check_session():
+        self.is_login = False
+        if self.__check_session():
+            self.is_login = True
+        else:
             self.logout()
-            self.is_login = False
+            if self.login(username, password):
+                self.is_login = True
 
     def contest_standings(self, contest_name: str):
         rank_url = f"{self.atcoder_url}/contests/{contest_name}/standings/json"
         headers = {"Content-Type": "application/json"}
         res = requests.get(rank_url, headers=headers, cookies=self.__load_cookies())
-        return res.json()
+        res_dict = res.json()
+        res_dict["ContestName"] = contest_name
+        return res_dict
 
     def user_history(self, username: str):
         history_url = f"{self.atcoder_url}/users/{username}/history/json"
@@ -30,8 +34,14 @@ class Fetcher:
         res = requests.get(history_url, headers=headers)
         return res.json()
 
+    @staticmethod
+    def contests_information():
+        atcoder_problems_url = "https://kenkoooo.com/atcoder/resources/contests.json"
+        res = requests.get(atcoder_problems_url)
+        return res.json()
+
     def login(self, username: str, password: str):
-        if not self.is_login:
+        if self.is_login:
             return False
 
         login_url = f"{self.atcoder_url}/login"
@@ -45,15 +55,6 @@ class Fetcher:
         self.__save_cookies(cookies)
         return True
 
-    def __check_session(self):
-        try:
-            settings_url = f"{self.atcoder_url}/settings"
-            r = requests.get(settings_url, cookies=self.__load_cookies())
-            return r.status_code == 200
-
-        except Exception:
-            return False
-
     def logout(self):
         if not self.is_login:
             return False
@@ -62,7 +63,9 @@ class Fetcher:
         r = client.get(self.atcoder_url)
         params = {"csrf_token": self.__extract_csrf_token(r.content)}
         client.post(logout_path, params=params)
-        os.remove(self.config_path + self.session_filename)
+        if os.path.exists(self.config_path + self.session_filename):
+            os.remove(self.config_path + self.session_filename)
+        self.is_login = False
         return True
 
     def __load_cookies(self):
@@ -74,6 +77,18 @@ class Fetcher:
             os.mkdir(self.config_path)
         with open(self.config_path + self.session_filename, "wb") as f:
             pickle.dump(cookies, f)
+
+    def __check_session(self):
+        try:
+            if not os.path.exists(self.config_path + self.session_filename):
+                return False
+            settings_url = f"{self.atcoder_url}/settings"
+
+            r = requests.get(settings_url, cookies=self.__load_cookies())
+            return r.status_code == 200
+
+        except Exception:
+            return False
 
     @staticmethod
     def __extract_csrf_token(content):
